Split off parser test case data from test module
- This is done in a similar manner and reason as what the lexer tests
  went through.
